requirements.txt
pony
gevent
asciimatics
lxml
feedparser
click
requests
grequests
pytest
beautifulsoup4
dateparser
python-dateutil